refactor : separate files for Go method, resource, and security
